feat: implement attach command
